+++ app/src/panels/inspector/system/SystemTabs.tsx
@@ -16,6 +16,7 @@ import { InitializerSection } from "./sections/InitializerSection";
 import { MegastructureSection } from "./sections/Megastructures";
 import { PlanetSection } from "./sections/Planets";
 import { ResourceSection } from "./sections/Resources";
+import { FeLinksSection } from "./sections/scenario/FeLinksSection";
 import { FeZoneSection } from "./sections/scenario/FeZoneSection";
 import { MarauderSection } from "./sections/scenario/MarauderSection";
 import { ScriptsTab, SCRIPTS_TAB_TITLE } from "./sections/scenario/ScriptsTab";
@@ -86,6 +87,7 @@ function ScenarioOverview({
       <OverviewHead detail={detail} />
       <SpawnPointSection system={system} />
       <FeZoneSection system={system} />
+      <FeLinksSection system={system} />
       <MarauderSection system={system} />
       <WormholePairSection system={system} />
       <InitializerSection system={system} spawn={planets === 0} />